+++ src/main/java/roomescape/controller/reservation/ReservationController.java
@@ -9,7 +9,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.util.UriComponentsBuilder;
-import roomescape.service.ReservationService;
+import roomescape.service.reservation.ReservationService;
 
 import java.net.URI;
 import java.util.List;
