+++ src/main/java/roomescape/controller/theme/ThemeController.java
@@ -10,7 +10,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.util.UriComponentsBuilder;
-import roomescape.service.ThemeService;
+import roomescape.service.theme.ThemeService;
 
 import java.net.URI;
 import java.util.List;
