+++ src/main/java/roomescape/controller/time/TimeController.java
@@ -10,7 +10,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.util.UriComponentsBuilder;
-import roomescape.service.TimeService;
+import roomescape.service.time.TimeService;
 
 import java.net.URI;
 import java.util.List;
