+++ src/main/java/roomescape/service/auth/CookieAuthorizationExtractor.java
@@ -2,7 +2,7 @@ package roomescape.service.auth;
 
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
-import roomescape.service.exception.TokenNotFoundException;
+import roomescape.service.auth.exception.TokenNotFoundException;
 
 public class CookieAuthorizationExtractor implements AuthorizationExtractor<String> {
     private static final String COOKIE_NAME = "token";
