+++ src/main/java/roomescape/service/reservation/exception/PreviousTimeException.java
@@ -1,4 +1,4 @@
-package roomescape.service.exception;
+package roomescape.service.reservation.exception;
 
 import roomescape.exception.BadRequestException;
