+++ src/main/java/roomescape/service/reservation/exception/ReservationDuplicatedException.java
@@ -1,4 +1,4 @@
-package roomescape.service.exception;
+package roomescape.service.reservation.exception;
 
 import roomescape.exception.ConflictException;
