+++ src/main/java/roomescape/service/theme/ThemeService.java
@@ -1,4 +1,4 @@
-package roomescape.service;
+package roomescape.service.theme;
 
 import org.springframework.stereotype.Service;
 import roomescape.controller.theme.CreateThemeRequest;
@@ -8,10 +8,10 @@ import roomescape.controller.theme.ThemeResponse;
 import roomescape.domain.Theme;
 import roomescape.repository.ReservationRepository;
 import roomescape.repository.ThemeRepository;
-import roomescape.service.exception.DaysLimitException;
-import roomescape.service.exception.RowsLimitException;
-import roomescape.service.exception.ThemeNotFoundException;
-import roomescape.service.exception.ThemeUsedException;
+import roomescape.service.theme.exception.DaysLimitException;
+import roomescape.service.theme.exception.RowsLimitException;
+import roomescape.service.theme.exception.ThemeNotFoundException;
+import roomescape.service.theme.exception.ThemeUsedException;
 
 import java.time.LocalDate;
 import java.util.List;
