+++ src/main/java/roomescape/service/theme/exception/DaysLimitException.java
@@ -1,9 +1,9 @@
-package roomescape.service.exception;
+package roomescape.service.theme.exception;
 
 import roomescape.exception.BadRequestException;
 
 public class DaysLimitException extends BadRequestException {
-    
+
     public DaysLimitException(String message) {
         super(message);
     }
