+++ src/main/java/roomescape/service/time/exception/TimeDuplicatedException.java
@@ -1,4 +1,4 @@
-package roomescape.service.exception;
+package roomescape.service.time.exception;
 
 import roomescape.exception.ConflictException;
