+++ src/main/java/roomescape/service/time/exception/TimeUsedException.java
@@ -1,4 +1,4 @@
-package roomescape.service.exception;
+package roomescape.service.time.exception;
 
 import roomescape.exception.BadRequestException;
