+++ src/test/java/roomescape/service/ThemeServiceTest.java
@@ -20,10 +20,11 @@ import roomescape.repository.H2ReservationTimeRepository;
 import roomescape.repository.H2ThemeRepository;
 import roomescape.repository.ReservationRepository;
 import roomescape.repository.ReservationTimeRepository;
-import roomescape.service.exception.DaysLimitException;
-import roomescape.service.exception.RowsLimitException;
-import roomescape.service.exception.ThemeNotFoundException;
-import roomescape.service.exception.ThemeUsedException;
+import roomescape.service.theme.ThemeService;
+import roomescape.service.theme.exception.DaysLimitException;
+import roomescape.service.theme.exception.RowsLimitException;
+import roomescape.service.theme.exception.ThemeNotFoundException;
+import roomescape.service.theme.exception.ThemeUsedException;
 
 import java.time.LocalDate;
 import java.util.ArrayList;
