+++ src/test/java/roomescape/service/TimeServiceTest.java
@@ -16,9 +16,10 @@ import roomescape.repository.H2ReservationTimeRepository;
 import roomescape.repository.H2ThemeRepository;
 import roomescape.repository.ReservationRepository;
 import roomescape.repository.ThemeRepository;
-import roomescape.service.exception.TimeDuplicatedException;
-import roomescape.service.exception.TimeNotFoundException;
-import roomescape.service.exception.TimeUsedException;
+import roomescape.service.time.TimeService;
+import roomescape.service.time.exception.TimeDuplicatedException;
+import roomescape.service.time.exception.TimeNotFoundException;
+import roomescape.service.time.exception.TimeUsedException;
 
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
